+++ src/api/services/CustomerService.js
@@ -1,4 +1,7 @@
+const responseApi = require('../utils/responseApi')
+
 const CustomerRepository = require('../repositories/CustomerRepository');
+const CustomerAttemptRepository = require('../repositories/CustomerAttemptRepository');
 
 module.exports = {
     async findAll() {
@@ -7,6 +10,18 @@ module.exports = {
     },
 
     async create(customer) {
+
+       // inicializar resposta de erro
+       responseApi.statusCode = 200
+       
+        const emailHasCustomer = await CustomerRepository.findCustomerByEmail(customer.email)
+        if(emailHasCustomer && emailHasCustomer.length > 0)
+        {
+            responseApi.statusCode = 404
+            responseApi.resp = false
+            responseApi.message = 'O \"email\" informado já está cadastrado.'            
+            return responseApi
+        }
 
         const customerResp = await CustomerRepository.create(customer)
 
@@ -22,6 +37,20 @@ module.exports = {
     },
     async destroy(id) {
 
-        await CustomerRepository.destroy(id)
+        responseApi.resp = false
+        responseApi.message = 'Não será possível excluir este \"cliente\".'
+
+        const customerHasAttempt = await CustomerAttemptRepository.findByCustomerId(id)
+
+        if (customerHasAttempt && customerHasAttempt.length === 0) {
+
+            await CustomerRepository.destroy(id)
+            responseApi.resp = true
+            responseApi.message = 'Excluído com sucesso.'
+
+        }
+
+        return responseApi
+
     },
 }
